fix column builder

+++ src/main/java/core/project/chess/infrastructure/dal/util/sql/Util.java
@@ -5,12 +5,7 @@ public class Util {
     private Util() {}
 
     static void deleteSurplusComa(StringBuilder query) {
-        if (query.charAt(query.length() - 1) == ',') {
-            query.deleteCharAt(query.length() - 1);
-        }
-
-        if (query.charAt(query.length() - 2) == ',') {
-            query.deleteCharAt(query.length() - 2);
-        }
+        if (query.charAt(query.length() - 1) == ',') query.deleteCharAt(query.length() - 1);
+        if (query.charAt(query.length() - 2) == ',') query.deleteCharAt(query.length() - 2);
     }
 }
